refactor: Changed logic to the negative
In the previous commit the user would use `let g:select_no_ignore_vcs =
1` to OPT IN and allow ignored files to be selected. With this commit,
the logic is changed so that ignored files are shown by default and the
user OPTS OUT of this behaviour with `let g:select_no_ignore_vcs = 0`.

The code around `autoload/select/def.vim` that concerns this feature was
also improved.

Updated `README.adoc` to account for the changes.

Updated `doc/select.txt` to account for the changes.